+++ clients/admin-ui/src/types/api/models/WorkerStats.ts
@@ -1,0 +1,11 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { QueueInfo } from "./QueueInfo";
+import type { WorkerInfo } from "./WorkerInfo";
+
+export type WorkerStats = {
+  queues: Record<string, QueueInfo>;
+  workers: Record<string, WorkerInfo>;
+};